refactor more preference page methods
